Extract codes to lib and util
