+++ lib/getComponents.js
