+++ lib/getSections.js
